Enhance emoji picker with usage stats
The change implements a feature to track and display the most frequently used emojis in the emoji picker, instead of the recent ones. This enhances user experience by providing them faster access to their preferred emojis.

- Updated emoji-picker and toolbar to utilize top used emojis.
- Introduced `useEmojiStats` hook for tracking and saving emoji usage.
- Implemented local storage persistence for emoji usage statistics.
- Adjusted components to use `useEmojiStats` for dynamic emoji suggestions.
