+++ apps/web/src/components/chat/message-composer.tsx
@@ -1,10 +1,9 @@
 import type { Editor } from "@tiptap/react"
-import { Attachment01, FaceSmile, Recording02 } from "@untitledui/icons"
-import { useEffect, useRef, useState } from "react"
-import { Button } from "~/components/base/buttons/button"
-import { ButtonUtility } from "~/components/base/buttons/button-utility"
+import { useRef, useState } from "react"
+
 import { useChat } from "~/providers/chat-provider"
-import { TextEditor, useEditorContext } from "../base/text-editor/text-editor"
+import { TextEditor } from "../base/text-editor/text-editor"
+import { MessageComposerActions } from "./message-composer-actions"
 
 export const MessageComposer = () => {
 	const { sendMessage, startTyping, stopTyping } = useChat()
@@ -61,37 +60,12 @@ export const MessageComposer = () => {
 						<TextEditor.Tooltip />
 
 						<div className="relative flex flex-col gap-2">
-							<TextEditor.Content
-								ref={textareaRef}
-							/>
-							<MessageActions onSubmit={handleSubmit} />
+							<TextEditor.Content ref={textareaRef} />
+							<MessageComposerActions onSubmit={handleSubmit} />
 						</div>
 					</>
 				)}
 			</TextEditor.Root>
-		</div>
-	)
-}
-
-const MessageActions = ({ onSubmit }: { onSubmit?: (editor: Editor) => Promise<void> }) => {
-	const editor = useEditorContext()
-
-	return (
-		<div className="absolute right-3.5 bottom-2 flex items-center gap-2">
-			<div className="flex items-center gap-0.5">
-				<ButtonUtility icon={Attachment01} size="xs" color="tertiary" />
-				<ButtonUtility icon={FaceSmile} size="xs" color="tertiary" />
-			</div>
-
-			<Button
-				size="sm"
-				color="link-color"
-				onClick={async () => {
-					await onSubmit?.(editor.editor)
-				}}
-			>
-				Send
-			</Button>
 		</div>
 	)
 }
